Overhaul the configuration system for the middleware
The goals of this commit is to allow configuring the middleware
without having to use env-vars which is useful for testing.

Second as part of this we now implement JWKs properly, that is using
exactly every key listed which is required for the proper functioning
of the system, when keys are being rolled.